[Starlight] Land pack fix (#41)
